refactor(events): split DetailsStep under the line cap (#544)
DetailsStep.svelte was 997 lines. Extract two cohesive field-group
sub-components (flat prefixed siblings, matching the TierForm* convention):

- DetailsStepTagsInput.svelte: the tag autocomplete combobox — its own
  debounced-suggestion state machine plus the tag chip list. Props: tags,
  onUpdate.
- DetailsStepMediaSection.svelte: the logo/cover-art accordion section with
  its own logoFile/coverArtFile state and org-fallback preview URLs. Props:
  logo/coverArt/organizationLogo/organizationCoverArt, isOpen, onToggle,
  onUpdateImages.

The `let description = $derived(...)` writable-derived (from #535 Task 1) stays
in DetailsStep's Basic section, untouched. DetailsStep 997 -> 648. Behavior
unchanged.
